Refactor: Repository TestCode Class 하나로 모으기
